Websocket proxy support
This allows websockets to be proxied through to the underlying app
alongside http connections. If the underlying app is reloaded (via a
http request) the connected websockets will be closed (and can be
re-opened by their clients if desired).

Note that traffic over the websocket connections is not monitored in
any way with this change, so the trigger for the code reload of the
underlying app will still need to be done via a standard http
request.